source/neural_network/data/requirements.txt
requests~=2.32.3
pillow~=11.0.0
